refactor(hook): move useStepper to hook folder
