+++ src/ui/component/stepper/stepper.tsx
@@ -2,8 +2,8 @@ import { TransitionGroup, CSSTransition } from 'react-transition-group'
 import { type FC } from 'react'
 import type { Optional } from '@/util/type'
 import { StepperProgress } from './stepperProgress/stepperProgress'
-import type { Step, UseStepper } from './useStepper'
-import { useStepper } from './useStepper'
+import type { Step, UseStepper } from '../../hook/useStepper'
+import { useStepper } from '@/ui/hook/useStepper'
 import { StepperActions } from './stepperActions/stepperActions'
 import './stepper.scss'
